requirements.txt
# ============================================================
# Proyecto: OSINT Manuelita Agroindustrial — Knowledge Base
# Versión Python: 3.10+
# Última actualización: 2026-04-18
# ============================================================

# --- Scraping Web ---
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.2.0
playwright>=1.45.0          # Solo si se necesita JS rendering (instalar: playwright install chromium)

# --- PDFs ---
pdfplumber==0.11.0
PyMuPDF>=1.25.1             # Alternativa robusta: import fitz (>=1.25 tiene wheels para Py3.13)
pytesseract==0.3.10         # Requiere Tesseract instalado en el sistema
Pillow>=10.4.0              # >=10.4 tiene wheels para Py3.13

# --- Noticias y Prensa ---
newspaper3k==0.2.8
feedparser==6.0.11

# --- YouTube Data API ---
google-api-python-client==2.120.0
yt-dlp==2024.3.10

# --- NLP y Entidades ---
spacy>=3.8.0                # >=3.8 soporta Python 3.13
# Instalar modelo español: python -m spacy download es_core_news_lg
langdetect==1.0.9

# --- Datos y Tablas ---
pandas>=2.2.2               # >=2.2.2 soporta numpy 2.x / Py3.13

# --- Markdown y Templates ---
python-frontmatter==1.1.0
jinja2==3.1.4
Markdown==3.6

# --- Configuración y Entorno ---
python-dotenv==1.0.1

# --- Logging y Retries ---
loguru==0.7.2
tenacity==8.2.3

# --- User Agents y Utilidades ---
fake-useragent==1.4.0
tqdm==4.66.2
python-slugify==8.0.4
dateparser==1.2.0

# --- Deduplicación ---
datasketch==1.6.4           # MinHash LSH para deduplicación semántica
